deploy_service/pyproject.toml
[project]
name = "week4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "mlflow==3.1.1",
    "scikit-learn>=1.7.0",
    "xgboost>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "google-cloud-storage>=2.10.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.4.1",
]
